handling file uploads, pdf generation and more on error handling

+++ controlers/admin.js
@@ -1,8 +1,7 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator/check');
-
-const mongoose = require("mongoose");
-const { returnError } = require('./../util/utilFunctions');
+const { returnError, deleteFile } = require('./../util/utilFunctions');
+const path = require('path');
 
 exports.getAddProduct = (req, res, next) => {
     return res.render('admin/edit-product', {
@@ -17,23 +16,35 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const description = req.body.description;
     const price = +req.body.price;
 
+    let validationErrors = {};
+    if (!image) {
+        validationErrors.image = 'Attached file is not an image';
+    }
+
     const errors = validationResult(req);
 
-    if (!errors.isEmpty()) {
+    if (!errors.isEmpty() || !image) {
+        validationErrors = {
+            ...validationErrors,
+            ...errors.array().reduce((map, value) => (map[value.param] = value.msg, map), {})
+        }
+        console.log("!errors.isEmpty() || !image");
         return res.render('admin/edit-product', {
             pageTitle: 'Edit product',
             path: '/admin/add-product',
             editing: false,
             buttonLabel: 'Update Product',
             formAction: '/admin/add-product',
-            oldInputs: { title, price, description, imageUrl },
-            validationErrors: errors.array().reduce((map, value) => (map[value.param] = value.msg, map), {})
+            oldInputs: { title, price, description },
+            validationErrors
         });
     }
+
+    const imageUrl = `/${image.path.replace(/\\/g, '/')}`;;
 
     const product = new Product({
         title: title,
@@ -44,7 +55,7 @@ exports.postAddProduct = (req, res, next) => {
     });
     product.save()
         .then(() => res.redirect('/admin/products'))
-        .catch(err => console.log(err));
+        .catch(err => returnError(err, next));
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -65,7 +76,7 @@ exports.getEditProduct = (req, res, next) => {
                 validationErrors: []
             });
         })
-        .catch(err => console.log(err));
+        .catch(err => returnError(err, next));
 };
 
 exports.postEditProduct = (req, res, next) => {
@@ -73,7 +84,7 @@ exports.postEditProduct = (req, res, next) => {
     const title = req.body.title;
     const price = req.body.price;
     const description = req.body.description;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const editMode = req.query.edit === 'true';
 
     const errors = validationResult(req);
@@ -85,7 +96,7 @@ exports.postEditProduct = (req, res, next) => {
             editing: editMode,
             buttonLabel: 'Update Product',
             formAction: '/admin/edit-product',
-            oldInputs: { _id: id, title, price, description, imageUrl },
+            oldInputs: { _id: id, title, price, description },
             validationErrors: errors.array().reduce((map, value) => (map[value.param] = value.msg, map), {})
         });
     }
@@ -99,12 +110,16 @@ exports.postEditProduct = (req, res, next) => {
             product.title = title;
             product.price = price;
             product.description = description;
-            product.imageUrl = imageUrl;
+
+            if (image) {
+                deleteFile(product.imageUrl);
+                product.imageUrl = `/${image.path.replace(/\\/g, '/')}`;
+            }
 
             return product.save().
                 then(() => res.redirect('/admin/products'));
         })
-        .catch(err => console.log(err));
+        .catch(err => returnError(err, next));
 };
 
 exports.getProducts = (req, res, next) => {
@@ -115,12 +130,21 @@ exports.getProducts = (req, res, next) => {
                 pageTitle: 'Admin Products',
                 path: '/admin/products'
             });
-        }).catch(err => console.log(err));
+        }).catch(err => returnError(err, next));
+
 };
 
 exports.postDeleteProduct = (req, res, next) => {
     const id = req.body.id;
-    Product.deleteOne({ _id: id, userId: req.user._id })
+
+    Product.findById(id)
+        .then(product => {
+            if (!product) {
+                return returnError('Product not found', next, 404);
+            }
+            deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: id, userId: req.user._id });
+        })
         .then(() => res.redirect('/admin/products'))
-        .catch(err => console.log(err));
+        .catch(err => returnError(err, next));
 };
